+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 5.4 [????.??.??]
 ----------------
 * Remove instances for `Option`, which was removed in `base-4.16`.
+* Add `Alt` instance for `Identity`.
 
 5.3.5 [2020.12.31]
 ------------------